Rename 'async' module to 'asyncUtil'
(`async` is a future reserved word)
